refactor(SCIM): rename {saml,scim}/SectionName.tsx to index.tsx
Same pattern as `general-tab/index.tsx` and the parent `sso/index.tsx`
— each feature folder owns one entry point at `index.tsx`. Lets the
parent import as `from './saml'` and `from './scim'` instead of
naming the file twice in the path.
